staging-to-master (#2)
# Cleanup
- removing of unused usings
- sorted usings 'silibum'-style
- shortened some methods
- misc

#Added EcoCompaniesMod checking
The EcoCompaniesMod is now detected and the `Company Legal Person` is now also in-jurisidication for the duration of any law based actions...

# Add option to disable internal transfers
Introduces the NoInternalTransfer config option to SmartTaxPlugin, allowing all actions to use external transfers instead of internal ones. Refactors transfer logic in TaxCard to use a new HandleTransfer method, enabling laws to see transfers at the cost of potential performance impact.